+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Fixes, improvements and extensions of config notifications [#744](https://github.com/rokwire/safer-illinois-app/issues/744).
 
 ## [2.12.4] - 2021-10-29
 ### Added